+++ authority-portal-backend/authority-portal-api/src/main/java/de/sovity/authorityportal/api/model/CreateConnectorStatusDto.java
@@ -1,0 +1,10 @@
+package de.sovity.authorityportal.api.model;
+
+import io.swagger.v3.oas.annotations.media.Schema;
+
+@Schema(description = "Possible status values for connector registration response", enumAsRef = true)
+public enum CreateConnectorStatusDto {
+    OK,
+    WARNING,
+    ERROR
+}